requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
pydantic==2.5.0
httpx==0.25.0
selenium==4.15.2
python-dotenv==1.0.0
